Add Options/Switches
A new module, "minimist", has been added for use in the parsing
options passed in commands. The passed arguments are then bound
to the relevant option function as its this argument.

Methods have been commented better in "index.js". Spacing has
been added to "tests/test.parser.js".

A new test to ensure that the `this` is bound with the parsed
options to the option function has been added.

The Dependency file, "package.json", has been changed to reflect
the new module dependency.